Separate extension for oembed converter
